feat: moved `utils` to `translatable_shared` and reordered files
