+++ translatable_shared/src/lib.rs
@@ -1,13 +1,3 @@
-mod languages;
-mod translations;
-
-/// Export all the structures in the common
-/// top-level namespace
-pub use crate::languages::{Language, LanguageIter, Similarities};
-pub use crate::translations::collection::TranslationNodeCollection;
-pub use crate::translations::node::{
-    TranslationNesting,
-    TranslationNode,
-    TranslationNodeError,
-    TranslationObject,
-};
+pub mod language;
+pub mod macros;
+pub mod translations;
