+++ translatable_shared/src/macros/errors.rs
@@ -22,6 +22,7 @@ where
 
 impl<T: Display> IntoCompileError for T {}
 
+#[macro_export]
 macro_rules! handle_macro_result {
     ($val:expr) => {{
         use $crate::utils::errors::IntoCompileError;
@@ -32,5 +33,3 @@ macro_rules! handle_macro_result {
         }
     }};
 }
-
-pub(crate) use handle_macro_result;
