+++ translatable_shared/src/macros/templating.rs
@@ -1,2 +1,1 @@
 
-
